Project submission
alphabetcount.c reads the document to count the frequency of alphabet characters (A-Z, a-z), creates a result.txt and prints the results into the file.

specialcharcount.c reads the document to count the frequency of special characters ( , . : ; !), creates a result.txt and prints the results into the file.

testalphabetcount.c and testspecialcharcount.c both prints the results.

added:
count.h, makefile, README, zero.zip